Add files via upload
This is a website made with django, therefore, to see and use it you must unzip it and download django "pip install django" in terminalif you have pip and a windows device.
You then have to cd (also in terminal) to whatever folder you downloaded the website to and to the folder and type "py manage.py runserver". the website code is quite messy (like most of my code) and there are surely a lot of useless stuff. My html isn't very good either but it does the job.